backend/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import uvicorn
from typing import List, Optional
import os
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

app = FastAPI(
    title="RAG API",
    description="基于RAG框架的智能问答系统API",
    version="1.0.0"
)

# 配置CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # 在生产环境中应该设置具体的源
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 健康检查端点
@app.get("/health")
async def health_check():
    return {"status": "healthy"}

# 文档处理接口
@app.post("/process")
async def process_document():
    # TODO: 实现文档处理逻辑
    pass

@app.post("/save")
async def save_document():
    # TODO: 实现文档保存逻辑
    pass

@app.get("/list-docs")
async def list_documents():
    # TODO: 实现文档列表获取逻辑
    pass

# 向量化接口
@app.post("/embed")
async def embed_document():
    # TODO: 实现文档向量化逻辑
    pass

@app.get("/list-embedded")
async def list_embedded_documents():
    # TODO: 实现向量化文档列表获取逻辑
    pass

# 检索接口
@app.post("/search")
async def search_documents():
    # TODO: 实现文档搜索逻辑
    pass

@app.get("/collections")
async def list_collections():
    # TODO: 实现集合列表获取逻辑
    pass

@app.get("/collections/{provider}")
async def get_collection(provider: str):
    # TODO: 实现特定提供商集合获取逻辑
    pass

# 生成接口
@app.post("/generate")
async def generate_response():
    # TODO: 实现回答生成逻辑
    pass

@app.get("/generation/models")
async def list_models():
    # TODO: 实现模型列表获取逻辑
    pass

if __name__ == "__main__":
    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True) 