backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-core>=2.33.2
python-dotenv==1.0.0
langchain-core>=0.3.60
langchain-community>=0.3.24
langchain-openai>=0.3.17
chromadb==0.4.18
sentence-transformers>=4.1.0
openai>=1.79.0
python-docx==1.0.1
PyPDF2==3.0.1
numpy==1.26.2
pandas==2.1.3
tqdm>=4.66.1
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1; platform_system != "Windows"
bcrypt==4.0.1; platform_system == "Windows" and python_version >= "3.8"
markdown==3.8
beautifulsoup4>=4.13.3
openpyxl==3.1.2
python-pptx>=1.0.2
striprtf==0.0.26
pillow>=10.4.0
psutil>=7.0.0
grpcio==1.63.2
protobuf>=6.31.0
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system == "Windows"
sqlalchemy>=2.0.31
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.27.0
langdetect==1.0.9
opentelemetry-proto==1.31.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
ruff==0.1.6 